+++ NEWS.md
@@ -31,6 +31,9 @@
 
 - (v 0.0.1.9004) add co-publication network to IOTN case study.
 
+- (v 0.0.1.9007) add project summary and award amount information
+  across programs.
+
 # grantpubcite v0.0.1
 
 - (v. 0.0.1) Initial version. Don't show case study code by default,